+++ pmd-kotlin/src/main/java/net/sourceforge/pmd/lang/kotlin/types/InternalApiBridge.java
@@ -4,6 +4,8 @@
 
 package net.sourceforge.pmd.lang.kotlin.types;
 
+import java.util.List;
+
 import net.sourceforge.pmd.annotation.InternalApi;
 import net.sourceforge.pmd.lang.kotlin.ast.KotlinNode;
 import net.sourceforge.pmd.lang.kotlin.ast.KotlinParser.KtKotlinFile;
@@ -15,7 +17,7 @@ import net.sourceforge.pmd.lang.kotlin.rule.internal.KotlinTypeAnalysisContext;
  *
  * <p><b>None of this is published API, and compatibility can be broken anytime!</b>
  *
- * @since 7.26.0
+ * @since 7.27.0
  * @internalApi None of this is published API, and compatibility can be broken anytime!
  */
 @InternalApi
@@ -31,6 +33,10 @@ public final class InternalApiBridge {
     /** @see KotlinNodeTypeData#setReturnTypeName(KotlinNode, String) */
     public static void setReturnTypeName(KotlinNode node, String returnTypeName) {
         KotlinNodeTypeData.setReturnTypeName(node, returnTypeName);
+    }
+
+    public static void setAnnotationFqNames(KotlinNode node, List<String> fqnList) {
+        KotlinNodeTypeData.setAnnotationFqNames(node, fqnList);
     }
 
     /** @see KotlinNodeTypeData#setTypeInfoAvailable(KtKotlinFile) */