+++ docs/pages/release_notes.md
@@ -27,7 +27,8 @@ This is a {{ site.pmd.release_type }} release.
 #### Kotlin type-aware analysis
 
 Kotlin now supports type-aware analysis via the `auxClasspath` language property (see [#6677](https://github.com/pmd/pmd/issues/6677)).
-Resolved type names, return types, and annotation FQNs are available through `KotlinNodeTypeData` for use in Java-based rules.
+Resolved type names, return types, and annotation FQNs are available through
+{%jdoc kotlin::lang.kotlin.types.KotlinNodeTypeData %} for use in Java-based rules.
 
 Note: Type data is not yet accessible in XPath rules or the PMD Rule Designer. This will be added in the next version.
 
@@ -62,6 +63,11 @@ Note: Type data is not yet accessible in XPath rules or the PMD Rule Designer. T
     deprecated `getClassLoader()` anymore.  
     Using ClassLoaders directly is discouraged, as it is unclear, if and when the ClassLoaders should be closed to release their resources.
     By just configuring the auxClasspath, PMD internally can deal with that.
+
+#### Experimental API
+* kotlin
+  * {%jdoc kotlin::lang.kotlin.types.KotlinNodeTypeData %}: Provides the initial API to access type information
+    on Kotlin AST nodes. It's part of the new Kotlin type-aware analysis.
 
 ### ✨️ Merged pull requests
 <!-- content will be automatically generated, see /do-release.sh -->